Rename some classes and files
